requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bitarray==3.4.2
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
cytoolz==1.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dune_client==1.7.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.6.2
hexbytes==1.3.1
idna==3.10
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
ndjson==0.3.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
parsimonious==0.10.0
propcache==0.3.1
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
six==1.17.0
toolz==1.0.0
tqdm==4.67.1
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250602
types-setuptools==80.9.0.20250529
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
web3==7.12.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
